pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Professional Proof Logger - Secure Evidence Platform"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask-login>=0.6.3",
    "flask>=3.1.1",
    "flask-sqlalchemy>=3.1.1",
    "flask-wtf>=1.2.0",
    "flask-limiter>=3.5.0",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "werkzeug>=3.1.3",
    "speechrecognition>=3.14.3",
    "pyaudio>=0.2.14",
    "pydub>=0.25.1",
    "opentimestamps-client>=0.7.0",
    "reportlab>=4.0.0",
    "pillow>=10.0.0",
    "qrcode[pil]>=7.0.0",
    "pyotp>=2.8.0",
    "dnspython>=2.0.0",
    "email-validator>=2.0.0",
    "dkimpy>=1.0.0",
    "bleach>=6.1.0",
    "python-magic>=0.4.27",
    "cryptography>=41.0.0",
    "redis>=5.0.0",
]
